fix load time problem

+++ app/Http/Controllers/PlanController.php
@@ -36,13 +36,13 @@ class PlanController extends Controller
     {
         $organizations = Organization::all();
         $categories = Plan::groupBy('category')->pluck('category');
-        return view('pages.plans.add', compact('organizations','categories'));
+        return view('pages.plans.add', compact('organizations', 'categories'));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      */
     public function store(Request $request)
     {
@@ -56,7 +56,7 @@ class PlanController extends Controller
         $performer->birthday = $performerInfo['birthday'];
 
         $performer->phone = $performerInfo['phone'];
-        if ($request->filled('second_number')){
+        if ($request->filled('second_number')) {
             $performer->second_number = $performerInfo['second_number'];
         }
         $performer->gender = $performerInfo['gender'];
@@ -69,7 +69,7 @@ class PlanController extends Controller
 //        $plan->tags = $planInfo['tags'];
         $plan->organization_id = $planInfo['organization'];
 //        $plan->distance = $planInfo['distance'];
-        $plan->start_date = shamsi2miladi('Y/m/d',fa2en($performerInfo['start_date']));
+        $plan->start_date = shamsi2miladi('Y/m/d', fa2en($performerInfo['start_date']));
         $plan->level = $planInfo['level'];
         $plan->address = $planInfo['address'];
         $plan->implement_method = $planInfo['implement_method'];
@@ -84,7 +84,7 @@ class PlanController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -95,7 +95,7 @@ class PlanController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
@@ -106,8 +106,8 @@ class PlanController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -118,7 +118,7 @@ class PlanController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
@@ -129,18 +129,18 @@ class PlanController extends Controller
     public function observes($planID)
     {
         $plan = Plan::find($planID);
-        if( !$plan )
+        if (!$plan)
             return redirect()->route('plans.index');
 
         $plan = $plan->load('organization');
 
-        $observes = $plan->load('employers', 'areaCity')->observes()->with(['performer', 'plan'])->get()->map( function ($observe){
+        $observes = $plan->load('employers', 'areaCity')->observes()->with(['performer', 'plan'])->get()->map(function ($observe) {
             $observe->observe_date = miladi2shamsi('Y/m/d', $observe->observe_date);
             return $observe;
-        } );
+        });
 
         $planImages = Image::with('document')->where('plan_id', $planID)->get();
-        $planDocuments = $planImages->map(function ($image){
+        $planDocuments = $planImages->map(function ($image) {
             return $image->document;
         });
 //        $planDocuments = $plan->documents()->with('image')->get();
@@ -159,13 +159,13 @@ class PlanController extends Controller
         $courses = $request->courses;
         $suggests = $request->suggests;
 
-        if( $courses ){
+        if ($courses) {
             $found_other = array_search('other', $courses);
-            if( $found_other )
+            if ($found_other)
                 unset($courses[$found_other]);
         }
 
-        if( $suggests ){
+        if ($suggests) {
             foreach ($suggests as $suggest) {
                 $suggestModel = new Suggest();
                 $suggestModel->title = $suggest;
@@ -186,34 +186,50 @@ class PlanController extends Controller
         $plan->distance = $request->distance;
         $plan->latitude = $request->latitude;
         $plan->longitude = $request->longitude;
-        if ($request->filled('description')){
+        if ($request->filled('description')) {
             $plan->description = $request->description;
 
         }
-        if ($request->filled('location_type')){
+        if ($request->filled('location_type')) {
             $plan->location_type = $request->location_type;
 
         }
-        if( $request->is_special == 'yes' ){
+        if ($request->is_special == 'yes') {
             $plan->is_special = true;
             $plan->special_reason = $request->special_reason;
         }
-        if( $request->is_exhibition == 'yes' ){
+        if ($request->is_exhibition == 'yes') {
             $plan->is_exhibition = true;
             $plan->exhibition_level = $request->exhibition_level;
             $plan->exhibition_desire = $request->exhibition_desire;
         }
-        if( $request->has_employment == 'yes' ){
+        if ($request->has_employment == 'yes') {
             $plan->has_employment = true;
         }
-        if( $request->filled('branch_id') ){
+        if ($request->filled('branch_id')) {
             $plan->branch_id = $request->branch_id;
         }
+        if ($request->filled('bank_id')) {
+            $plan->bank_id = $request->bank_id;
+        }
+
+
+        if ($request->filled('installment')) {
+            $plan->installment = $request->installment;
+        }
+        if ($request->filled('loan_amount')) {
+            $plan->loan_amount = $request->loan_amount;
+        }
+        if ($request->filled('loan_time')) {
+            $plan->loan_time = shamsi2miladi('Y/m/d', fa2en($request->loan_time))->format('Y/m/d');
+//            dd($plan->loan_time);
+        }
+
         $plan->save();
 
         $observe = Observe::where('plan_id', $plan->id)->whereDate('observe_date', $stringDate)->first();
         $found = true;
-        if( !$observe ){
+        if (!$observe) {
             $observe = new Observe();
             $found = false;
         }
@@ -224,23 +240,26 @@ class PlanController extends Controller
         $observe->observe_date = $date;
         $observe->monthly_charge = $request->monthly_charge;
         $observe->monthly_income = $request->monthly_income;
-        $observe->net_worth  = $request->net_worth;
+        $observe->net_worth = $request->net_worth;
 
         $observe->save();
-        foreach($request->problems as $problem){
-            if(!ObserveProblem::where('observe_id',$observe->id)->where('problem_id',$problem)->first()){
-                ObserveProblem::create([
-                    'observe_id'=>$observe->id,
-                    'problem_id'=>$problem
-                ]);
+        if ($request->has('problems')) {
+
+            foreach ($request->problems as $problem) {
+                if (!ObserveProblem::where('observe_id', $observe->id)->where('problem_id', $problem)->first()) {
+                    ObserveProblem::create([
+                        'observe_id' => $observe->id,
+                        'problem_id' => $problem
+                    ]);
+                }
             }
 
+            ObserveProblem::where('observe_id', $observe->id)->whereNotIn('problem_id', $request->problems)->delete();
         }
-        ObserveProblem::where('observe_id',$observe->id)->whereNotIn('problem_id',$request->problems)->delete();
         foreach ($request->observe_files as $item) {
             $documentID = $item['document'];
             $document = $plan->documents()->find($documentID);
-            if(isset($item['file']) ){
+            if (isset($item['file'])) {
                 $image = new Image();
                 $image->url = upload($item['file']);
                 $image->document_id = $documentID;
@@ -256,13 +275,13 @@ class PlanController extends Controller
         $performer->courses()->sync($courses);
         $performer->save();
 
-        if( $request->employers ){
-            foreach ($request->employers as $emp){
+        if ($request->employers) {
+            foreach ($request->employers as $emp) {
                 $employer = Employer::where([
                     ['nationalityCode', $emp['nationalityCode']],
                     ['plan_id', $planID]
                 ])->first();
-                if( !$employer )
+                if (!$employer)
                     $employer = new Employer();
 
                 $employer->nationalityCode = $emp['nationalityCode'];
@@ -284,5 +303,15 @@ class PlanController extends Controller
     {
         $data = $request->input();
         return view('pages.observes.done', compact('data'));
+    }
+
+    public function holdPlan(Request $request, $id)
+    {
+        $plan = Plan::where('id', $id)->first();
+        $plan->update([
+            'on_hold' => $request->on_hold == 1,
+            'hold_reason' => $request->hold_reason ?? " "
+        ]);
+        return redirect()->back();
     }
 }
